Redirect eanalysis hook via typeclass
The hook should be composable.
Like, one analysis for int ranges+one analysis for sets
gives an analysis for both.

The hook allows analyses to rewrite the graph.
For `x has range 1...1`, add Constant 1 to the class.

To make this composable, we need to make views so that the range
analysis only interacts with the range data of the egraph, hence the
typeclass